refactor: Make both classes use a common QEMU interface
Set up a AbstractCXL and a concrete QEMU CXL class that handles the low level details of interfacing with the QEMU device driver. Then, when we actually get a more faithful cxl emulation interface, we shud be able to swap classes easily. Note that this commit is not 100% done, we still leak some of our abstractions for the sake of making it work. Will clean up and improve next
