+++ qemu_share/clientlib/build_client.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-g++ -std=c++17 -g rpcclient.cpp client_main.cpp -o rpcclient
+g++ -std=c++17 -g ../includes/qemu_cxl_connector.cpp rpcclient.cpp client_main.cpp -o rpcclient
