+++ qemu_share/clientlib/rpcclient.hpp
@@ -7,6 +7,7 @@
 
 #include "../includes/cxl_switch_ipc.h"
 #include "../includes/ioctl_defs.h"
+#include "../includes/qemu_cxl_connector.hpp"
 
 namespace diancie {
 
@@ -19,20 +20,18 @@ struct ChannelInfo {
 
 // The client library manages the low-level interactions for the user program.
 // 1. Discover service and get allocated channel
-class DiancieClient {
+class DiancieClient : protected QEMUCXLConnector {
 public:
   DiancieClient() = delete;
   DiancieClient(const std::string& device_path, const std::string& service_name, const std::string& instance_id);
   ~DiancieClient();
   
-  // Generic interface methods
-  void write_u64(uint64_t offset, uint64_t value);
-  uint64_t read_u64(uint64_t offset);
+  // For testing purposes, not final
+  void client_write_u64(uint64_t offset, uint64_t value);
+  uint64_t client_read_u64(uint64_t offset);
+
 
 private:
-  std::string device_path_;
-  int device_fd_ = -1;
-
   // For logging purposes
   std::string service_name_;
   std::string instance_id_;
@@ -40,35 +39,10 @@ private:
   // For release
   uint64_t channel_id_;
   
-  void* bar0_base_ = nullptr;  // mailbox
-  size_t bar0_size_ = 0;
-  
-  void* bar1_base_ = nullptr;  // control registers
-  size_t bar1_size_ = 0;
-  
-  void* bar2_base_ = nullptr;  // data window
-  size_t bar2_size_ = 0;
-
-  int eventfd_cmd_ready_ = -1;
-
   // Methods that are called by ctor and dtor
   // Find a service via the FM and request a comms channel
   std::optional<ChannelInfo> request_channel(const std::string& service_name, const std::string& instance_id);
-  // Configure local QEMU device's BAR2 to map to channel
-  // TODO: Think about what shud be field later
-  bool set_memory_window(uint64_t offset, uint64_t size);
   bool release_channel();
-
-  bool setup_eventfd(int& efd, unsigned int ioctl_cmd);
-  bool wait_for_command_response(int timeout_ms);
-  uint32_t get_command_status();
-  bool send_command(const void* req, size_t size);
-
-  static constexpr off_t BAR0_MMAP_OFFSET = 0 * 4096; // MMAP_OFFSET_PGOFF_BAR0
-  static constexpr off_t BAR1_MMAP_OFFSET = 1 * 4096; // MMAP_OFFSET_PGOFF_BAR1
-  static constexpr off_t BAR2_MMAP_OFFSET = 2 * 4096; // MMAP_OFFSET_PGOFF_BAR1
-  static constexpr size_t DEFAULT_BAR_SIZE = 4096;
-  static constexpr size_t DEFAULT_BAR2_SIZE = 256 * 1024 * 1024;
 };
 
 }
