+++ qemu_share/includes/a_cxl_connector.hpp
@@ -1,0 +1,28 @@
+#ifndef A_CXL_CONNECTOR_HPP
+#define A_CXL_CONNECTOR_HPP
+
+#include <cstddef>
+#include <stdint.h>
+namespace diancie {
+
+/// This abstract class is responsible for all low-level interactions with
+/// the underlying CXL Switch device. It will make it possible to easily
+/// (hopefully) transition between our current QEMU emulation for correctness
+/// verification and a more faithful (but challenging to set up) configuration
+class AbstractCXLConnector {
+public:
+  AbstractCXLConnector() = default;
+  virtual ~AbstractCXLConnector() = default;
+protected:
+  virtual bool wait_for_command_response(int timeout_ms) = 0;
+  virtual uint32_t get_notification_status() = 0;
+  virtual bool send_command(const void* req, size_t size) = 0;
+  virtual bool recv_response(void* resp, size_t size) = 0;
+  // In the real deployment, must configure the window (if CXL: it wud be the logical view)
+  virtual bool set_memory_window(uint64_t offset, uint64_t size) = 0;
+};
+
+}; // namespace diancie
+
+
+#endif
