+++ qemu_share/serverlib/build_server.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-g++ -std=c++17 rpcserver.cpp concurrent_server.cpp -o ccserver
+g++ -std=c++17 ../includes/qemu_cxl_connector.cpp rpcserver.cpp concurrent_server.cpp -o ccserver
